refactor: merge helpers
