+++ src/commands/basic/component.js
@@ -6,11 +6,13 @@ import inquirer from 'inquirer';
 import showBanner from 'node-banner';
 import path from 'path';
 
-import appData from '../../utils/projectConfig';
-import { checkIfConfigFileExists } from '../../utils/messages';
 import exec from '../../utils/exec';
+import {
+  checkIfConfigFileExists,
+  fetchProjectConfig,
+  readFileContent,
+} from '../../utils/helpers';
 import { validateInput } from '../../utils/validate';
-import readFileContent from '../../utils/helpers';
 
 /**
  * Converts a given string into lower camel case
@@ -28,7 +30,7 @@ const toLowerCamelCase = (str) =>
  * @returns {Promise<void>}
  */
 
-const generateComponent = async () => {
+export default async () => {
   await showBanner('MEVN CLI', 'Light speed setup for MEVN stack based apps.');
   checkIfConfigFileExists();
 
@@ -45,7 +47,7 @@ const generateComponent = async () => {
   ]);
 
   // Fetch information specific to the project
-  const projectConfig = appData();
+  const projectConfig = fetchProjectConfig();
   const { template, isConfigured } = projectConfig;
 
   const { componentType } = await inquirer.prompt([
@@ -187,5 +189,3 @@ const generateComponent = async () => {
     },
   );
 };
-
-module.exports = generateComponent;
