+++ src/commands/basic/generate.js
@@ -5,13 +5,15 @@ import inquirer from 'inquirer';
 import path from 'path';
 import showBanner from 'node-banner';
 
-import appData from '../../utils/projectConfig';
-import { checkIfConfigFileExists } from '../../utils/messages';
-import copyDirSync from '../../utils/fs';
 import exec from '../../utils/exec';
 import generateComponent from './component';
+import {
+  checkIfConfigFileExists,
+  copyDirSync,
+  fetchProjectConfig,
+  readFileContent,
+} from '../../utils/helpers';
 import { validateInput } from '../../utils/validate';
-import readFileContent from '../../utils/helpers';
 
 // Holds reference to the path where the boilerplate files reside
 const templatePath = path.join(__dirname, '..', '..', 'templates');
@@ -34,7 +36,7 @@ const createDir = (dir) => {
  * @returns {Promise<void>}
  */
 
-const generateFile = async () => {
+export default async () => {
   await showBanner('MEVN CLI', 'Light speed setup for MEVN stack based apps.');
   checkIfConfigFileExists();
 
@@ -55,70 +57,72 @@ const generateFile = async () => {
   ]);
 
   // Fetch boilerplate template used from .mevnrc
-  const projectConfig = appData();
+  const projectConfig = fetchProjectConfig();
   const { template, isConfigured } = projectConfig;
 
+  // Generate client side component
   if (type.includes('Component')) {
     return generateComponent();
+  }
+
+  // Generate CRUD boilerplate for the server
+  if (template === 'GraphQL') {
+    // Create graphql-schema directory
+    createDir('graphql');
+
+    // Create models directory
+    createDir('models');
   } else {
-    if (template === 'GraphQL') {
-      // Create graphql-schema directory
-      createDir('graphql');
+    // Set up routes for CRUD functionality
+    const routesFilePath = path.join('server', 'routes', 'api.js');
+    fs.writeFileSync(
+      routesFilePath,
+      fs.readFileSync(path.join(templatePath, 'routes', 'index.js')),
+    );
 
-      // Create models directory
-      createDir('models');
-    } else {
-      // Set up routes for CRUD functionality
-      const routesFilePath = path.join('server', 'routes', 'api.js');
-      fs.writeFileSync(
-        routesFilePath,
-        fs.readFileSync(path.join(templatePath, 'routes', 'index.js')),
-      );
+    // Create controllers directory
+    createDir('controllers');
 
-      // Create controllers directory
-      createDir('controllers');
+    // Create models directory
+    createDir('models');
+  }
 
-      // Create models directory
-      createDir('models');
-    }
+  // Create .env file
+  const { uri } = await inquirer.prompt([
+    {
+      type: 'input',
+      name: 'uri',
+      message: 'Please provide the MongoDB URI path',
+      default: 'mongodb://localhost:27017',
+      validate: validateInput,
+    },
+  ]);
+  fs.writeFileSync(path.join('server', '.env'), `DB_URL=${uri}/userdb`);
 
-    // Create .env file
-    const { uri } = await inquirer.prompt([
-      {
-        type: 'input',
-        name: 'uri',
-        message: 'Please provide the MongoDB URI path',
-        default: 'mongodb://localhost:27017',
-        validate: validateInput,
-      },
-    ]);
-    fs.writeFileSync(path.join('server', '.env'), `DB_URL=${uri}/userdb`);
-
-    if (!isConfigured.server) {
-      // Installing dependencies
-      await exec(
-        'npm install',
-        'Installing dependencies',
-        'Dependencies were successfully installed',
-        {
-          cwd: 'server',
-        },
-      );
-      // .mevnrc
-      projectConfig.isConfigured.server = true;
-      fs.writeFileSync('.mevnrc', JSON.stringify(projectConfig, null, 2));
-    }
-
-    // Install mongoose ORM
+  if (!isConfigured.server) {
+    // Installing dependencies
     await exec(
-      'npm install --save mongoose',
-      'Installing mongoose ORM. Hold on',
-      `You're all set to interact with the DB`,
+      'npm install',
+      'Installing dependencies',
+      'Dependencies were successfully installed',
       {
         cwd: 'server',
       },
     );
+    // .mevnrc
+    projectConfig.isConfigured.server = true;
+    fs.writeFileSync('.mevnrc', JSON.stringify(projectConfig, null, 2));
   }
+
+  // Install mongoose ORM
+  await exec(
+    'npm install --save mongoose',
+    'Installing mongoose ORM. Hold on',
+    `You're all set to interact with the DB`,
+    {
+      cwd: 'server',
+    },
+  );
 
   copyDirSync(path.join(templatePath, 'helpers'), 'server');
 
@@ -139,5 +143,3 @@ const generateFile = async () => {
 
   fs.writeFileSync(serverFilePath, serverFileContent.join('\n'));
 };
-
-module.exports = generateFile;
