+++ src/commands/deploy/heroku.js
@@ -5,9 +5,8 @@ import execa from 'execa';
 import fs from 'fs';
 import inquirer from 'inquirer';
 import path from 'path';
-import readFileContent from '../../utils/helpers';
 
-import appData from '../../utils/projectConfig';
+import { fetchProjectConfig, readFileContent } from '../../utils/helpers';
 import { validateInput, validateInstallation } from '../../utils/validate';
 
 /**
@@ -86,11 +85,11 @@ const setConfigVar = async (configVar, dir, value) => {
  * @returns {Promise<void>}
  */
 
-const deployToHeroku = async (templateDir) => {
+export default async (templateDir) => {
   await validateInstallation('heroku');
   await validateInstallation('git help -g');
 
-  const projectConfig = appData();
+  const projectConfig = fetchProjectConfig();
   const { template } = projectConfig;
 
   if (!fs.existsSync(path.join(templateDir, '.git'))) {
@@ -182,8 +181,11 @@ const deployToHeroku = async (templateDir) => {
       const serverFilePath = path.join('client', 'server.js');
       if (!fs.existsSync(serverFilePath)) {
         fs.writeFileSync(serverFilePath, starterSource.join('\n'));
+
+        // Add preinstall script
         pkgJson.scripts['preinstall'] =
           'npm install --save express serve-static';
+
         fs.writeFileSync(
           './client/package.json',
           JSON.stringify(pkgJson, null, 2),
@@ -201,5 +203,3 @@ const deployToHeroku = async (templateDir) => {
   });
   await execa.command('heroku open', { cwd: templateDir });
 };
-
-module.exports = deployToHeroku;
