+++ src/utils/constants.js
@@ -2,12 +2,6 @@
 
 // Detecting host OS.
 
-const isWin = process.platform === 'win32';
-const isLinux = process.platform === 'linux';
-const isMac = process.platform === 'darwin';
-
-module.exports = {
-  isWin,
-  isLinux,
-  isMac,
-};
+export const isWin = process.platform === 'win32';
+export const isLinux = process.platform === 'linux';
+export const isMac = process.platform === 'darwin';
